Event page

+++ laravel5-angular-material-starter/angular/app/components/eventContent/eventContent.component.js
@@ -5,7 +5,7 @@ class EventContentController{
     /**
      * Injection des dependances necessaires dans le constructeur, ici, des acces API
      */
-    constructor(EventService, UserService, CategoryService, ToastService, DialogService, $state, $sce, $filter, $auth, $scope){
+    constructor(EventService, UserService, CategoryService, ToastService, DialogService, $state, $sce, $filter, $auth){
         'ngInject';
 
         this.EventService = EventService;
@@ -17,7 +17,6 @@ class EventContentController{
         this.$sce = $sce;
         this.$filter = $filter;
         this.$auth = $auth;
-        this.$scope = $scope;
         
         this.event = null;
         this.place = null;
@@ -28,9 +27,14 @@ class EventContentController{
         this.directions = null;
         this.user = null;
         this.invitations = null;
+        this.weather = null;
         
         this.visibleDirections = false;
+        this.visibleInvitations = false;
         this.userParticipation = false;
+        
+        this.organizersToAdd = [];
+        this.guestsToAdd = [];
         
         this.alreadyExistingEventDialog = function() {
           this.DialogService.fromTemplate('alreadyExistingEventDialog');
@@ -57,6 +61,7 @@ class EventContentController{
                 this.getCategory(dataCategory);
                 this.getPlace(dataPlace);
                 this.getDirections(dataPlace);
+                this.getWeather(dataPlace);
             },
             (responseError) => {
                 this.$state.go('app.landing');
@@ -119,6 +124,8 @@ class EventContentController{
                     this.comments = responseSuccess.data.comments;
                 else
                     this.comments = [];
+                
+                console.log(responseSuccess);
             },
             (responseSuccess) => {
                 //On n'a pas trouve les commentaires
@@ -177,7 +184,7 @@ class EventContentController{
             );
         }
     }
-    
+        
     /**
      * Methode permettant de recuperer les invites
      */
@@ -195,6 +202,178 @@ class EventContentController{
                 console.log('Erreur');
             }
         );     
+    }
+    
+    /**
+     * Methode permettant de recuperer une liste filtree d'utilisateurs
+     */
+    getFilteredUserList(selector) {
+        return this.UserService.findAll().then(
+            (success) => {
+                let users = [];
+                
+                success.forEach((user) => {
+                    let combinedName = user.firstName + ' ' + user.name;
+                    
+                    if(combinedName.includes(selector))
+                        users.push({id: user.id, name: user.firstName + " " + user.name, email: user.email });
+                });
+
+                return users;
+            },
+            (error) => {
+                //En cas d'erreur
+                return ['Impossible de charger les utilisateurs...'];
+            }
+        );
+    }
+    
+    /**
+     * Methode permettant de recuperer la meteo
+     */
+    getWeather(data) {
+        this.EventService.getWeather(data).then(
+            (responseSuccess) => {
+                //On a recupere la meteo
+                let dateOptimale = new Date();
+                let heure = dateOptimale.getHours();
+                
+                if(heure >= 0 && heure <= 2)
+                   dateOptimale.setHours(2, 0, 0);
+                else if(heure > 2 && heure <= 5)
+                    dateOptimale.setHours(5, 0, 0);
+                else if(heure > 5 && heure <= 8)
+                    dateOptimale.setHours(8, 0, 0);
+                else if(heure > 5 && heure <= 11)
+                    dateOptimale.setHours(11, 0, 0);
+                else if(heure > 5 && heure <= 14)
+                    dateOptimale.setHours(14, 0, 0);
+                else if(heure > 5 && heure <= 17)
+                    dateOptimale.setHours(17, 0, 0);
+                else if(heure > 5 && heure <= 20)
+                    dateOptimale.setHours(20, 0, 0);
+                else if(heure > 5 && heure <= 23)
+                    dateOptimale.setHours(23, 0, 0);
+                
+                let indexTableau = this.$filter('date')(dateOptimale, "yyyy-MM-dd HH:mm:ss");
+                
+                this.weather = responseSuccess[indexTableau];
+            },
+            (responseError) => {
+                //On n'a pas reussi a recuperer la metot
+                console.log('Erreur');
+            }
+        );
+    }
+    
+    /**
+     * Transforme les donnees meteo en une chaine de caracteres utile
+     */
+    textWeather() {
+        if(this.weather !== null) {
+            let temperature = Math.round(this.weather.temperature['sol'] - 273);
+            let vent = Math.round(this.weather.vent_moyen['10m']);
+            let pluie = Math.round(this.weather.pluie);
+            
+            let avisMeteo = "";
+            
+            if(temperature < 0)
+                avisMeteo += "Il va faire très froid. La météo prévoit " + temperature + "°C.<br>";
+            else if(temperature >= 0 && temperature <= 10)
+                avisMeteo += "Il va faire froid. La météo prévoit " + temperature + "°C.<br>";
+            else if(temperature > 10 && temperature <= 20)
+                avisMeteo += "La température sera moyenne. La météo prévoit " + temperature + "°C.<br>";
+            else if(temperature > 20 && temperature <= 30)
+                avisMeteo += "De belles températures en perspective ! La météo prévoit " + temperature + "°C.<br>";
+            else
+                avisMeteo += "De grosses chaleurs sont attendues. La météo envisage " + temperature + "°C.<br>";
+            
+            if(pluie > 0 && pluie < 5)
+                avisMeteo += " Quelques averses sont prévues. Pensez à prendre un parapluie !<br>";
+            else if(pluie > 5)
+                avisMeteo += " Des pluies relativement conséquentes sont prévues. Pensez à prendre un parapluie !<br>";
+            else
+                avisMeteo += " Aucune précipitation n'est prévue. Laissez votre parapluie à la maison !<br>";
+            
+            if(vent >= 15 && vent <= 30)
+                avisMeteo += " Un peu de vent à prévoir avec " + vent + " km/h en moyenne.";
+            else if(vent > 30)
+                avisMeteo += " Beaucoup de vent à prévoir avec " + vent + " km/h en moyenne.";
+            
+            return avisMeteo;
+        } else return "Les données météorologiques sont indisponibles pour le moment.";
+    }
+    
+    /**
+     * Methode permettant d'ajouter des organisateurs
+     */
+    addOrganizers() {
+        let organizersToAddLength = this.organizersToAdd.length;
+        let increment = 0;
+        
+        this.organizersToAdd.forEach((organizer) => {
+            var data = {
+                id: this.event.id,
+                idUser: organizer.id
+            }
+            
+            this.EventService.addOrganizer(data).then(
+                (success) => {
+                    increment++;
+                    
+                    if(increment === organizersToAddLength) {
+                        this.organizersToAdd = [];
+                        this.getOrganizers({ id: this.event.id });
+                    }        
+                }, 
+                (error) => {
+                    increment++;
+                    
+                    if(increment === organizersToAddLength) {
+                        this.organizersToAdd = [];
+                        this.getOrganizers({ id: this.event.id });
+                    }
+                    this.ToastService.error('Certains utilisateurs étaient déjà des organisateurs.');
+                }
+            ); 
+        });
+    }
+    
+    /**
+     * Methode permettant d'ajouter des invites
+     */
+    addInvitations() {
+        let guestsToAddLength = this.guestsToAdd.length;
+        let increment = 0;
+        
+        this.guestsToAdd.forEach((guest) => {
+            var data = {
+                id: this.event.id,
+                idUser: guest.id
+            }
+            
+            this.EventService.addInvitation(data).then(
+                (success) => {
+                    increment++;
+                    
+                    if(increment === guestsToAddLength) {
+                        this.guestsToAdd = [];
+                        this.getGuests({ id: this.event.id });
+                        this.visibleInvitations = false;
+                    }        
+                }, 
+                (error) => {
+                    increment++;
+                    
+                    if(increment === guestsToAddLength) {
+                        this.guestsToAdd = [];
+                        this.getGuests({ id: this.event.id });
+                        this.visibleInvitations = false;
+                    }
+                    this.ToastService.error('Certains utilisateurs étaient déjà invités.');
+                }
+            ); 
+        }); 
     }
     
     /**
@@ -291,16 +470,20 @@ class EventContentController{
                 id: this.event.id
             }    
     
-            this.EventService.deleteParticipant(data).then(
-                (success) => {
-                    this.ToastService.show(success.data);
-                    this.userParticipation = false;
-                    this.getParticipants(data);
-                },
-                (error) => {
-                    console.log(error);
-                }
-            );   
+            if(!this.isOrganizer()) {
+                this.EventService.deleteParticipant(data).then(
+                    (success) => {
+                        this.ToastService.show(success.data);
+                        this.userParticipation = false;
+                        this.getParticipants(data);
+                    },
+                    (error) => {
+                        console.log(error);
+                    }
+                );         
+            } else {
+              this.DialogService.alert('Vous êtes organisateur de cet événement.', 'Vous êtes organisateur de cet événement. Par conséquent, vous ne pouvez pas retirer votre participation. Si vous êtes plusieurs à organiser cet événement, retirez-vous de la liste des organisateurs pour être en mesure de retirer votre participation.');  
+            }  
         } else {
             this.ToastService.error('Vous devez être authentifié pour retirer votre participation à un événement.');
         } 
@@ -313,14 +496,14 @@ class EventContentController{
         if(this.$auth.isAuthenticated()) {
             var data = {
                 id: this.event.id,
-                comment: this.$scope.commentForm
+                comment: this.commentForm
             }
             
             this.EventService.addComment(data).then(
                 (success) => {
                     this.ToastService.show(success.data);
                     this.getComments(data);
-                    this.$scope.commentForm = "";
+                    this.commentForm = "";
                 }, 
                 (error) => {
                     console.log(error);
@@ -352,6 +535,21 @@ class EventContentController{
         } else this.userParticipation = false;
     }
     
+    isOrganizer() {
+        let userIsOrganizer = false;
+        
+        if(this.$auth.isAuthenticated()) {
+            if(this.organizers !== null && this.user !== null) {
+                this.organizers.forEach((organizer) => {
+                    if(organizer.id === this.user.id)
+                        userIsOrganizer = true;
+                })
+            }
+        }
+        
+        return userIsOrganizer;
+    }
+        
     /**
      * Lorsque quelque chose change
      */
